backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
pyswip==0.2.10
anthropic==0.34.2
httpx==0.27.2
google-cloud-storage==2.18.2
pytest==8.3.3
